first clean up
- created separate function holder file in the tools
- clean up console logs

- variable should be renamed
- should remove redundant lines
- should comment
- should add more edge handlings
